fix: align SDK with OpenAPI spec — add missing fields, fix types, and clean up
Reviewed all 30 services (114 endpoints) against openapi.json and fixed:
- Add missing metadata field to APIError.data in session error events
- Add MarshalJSON to McpAddConfigOAuthParam for correct serialization
- Fix SessionListParams.Limit type from int64 to float64 per spec
- Reorder session.error union variants to match spec ordering
- Fix doc comment URL typo (/docs/agent -> /docs/agents)
- Remove spurious URLQuery() on AuthSetParams/AuthRemoveParams
- Document known unimplemented GET /pty/{ptyID}/connect (WebSocket)

+++ session.go
@@ -3673,7 +3673,7 @@ type SessionListParams struct {
 	Roots     param.Field[bool]    `query:"roots"`
 	Start     param.Field[float64] `query:"start"`
 	Search    param.Field[string]  `query:"search"`
-	Limit     param.Field[int64]   `query:"limit"`
+	Limit     param.Field[float64] `query:"limit"`
 }
 
 // URLQuery serializes [SessionListParams]'s query parameters as `url.Values`.
